Added Tricord (with backups) + Build Optimizations
Took the base unfinished/broke GiGi Tricord and combined it with some of the work I've done on the GiGi to fix/finish it and put my own changes to it. Currently it is still being worked on. The "Backup" files are because I worked on this offline, also it just makes it easier for me to look at older versions on hand.

Additionally, I made some optimizations to the builder. I removed a warning which has significantly sped up build times and performance and removing clipping problems. None of my changes seem to have any side effects to the building process that I can see so far.