Builder Optimizations
I made some changes to DIYBuilder.scad that significantly improved performance. MiniWarp.scad was used to test the changes.
The first major change was to the hadamard() function. OpenSCAD was showing warnings that it "len(a)" was undefined, to fix this, I added an additional conditional operator that checks if "a" is a list before sending it to len(a). This one change significantly improved build times, bringing the total render time from up to over a minute, down to less then a second. This change doesn't effect the outputted model in any way that I noticed.
The second change was commenting out some calls to BuildColumn() at the end of BuildKeyWells(). This didn't increase performance too significantly, however, it did fix the issue of Z-Fighting in the keywells. This also doesn't effect the outputted model in any way that I noticed but I left the code commented out just in case.