Tricord Alterations, One Star Variation, Renamed my layouts, Removed MiniWarp and backups, Rerendered GiGI
Altered the "Tricord" layout so that the pinky keys layout so they match the layout of the rest of the keys in the "Main Row". The change isnt really noticeable, it just unifies things more, may change back if it doesnt work. Additionally, I tried to clean up some of the comments and spacing throughout the files but sort of gave up at a certain point as I would have to do too much by hand and they gained positives arent that much as the file is readable regardless.

I created a spin on the Tricord layout called the One Star. The layout removes one of the "*" index keys and moves the remaining one to the center, using the similar layout of the Epigrammatic as a guide. This key is meant to be used with a 1.25u or 1.5u key to follow the layout of a steno machine closer. May need some adjustments and fixes but it was just a super quick, 30 second alteration.

Removed the Miniwarp as I realized that I hadn't actually altered the layout, just changed the tenting. Additionally, I removed my backup Tricord files.

I removed all of the "_Smeeba" Suffix after my files as they are already in their own file so it isnt really needed.

GiGi STL rerendered using my altered builder, resulting in a smaller file size.